refactors calibration to nested behavior Class in ArmArbitrator, adds Stop behavior nested in ArmArbitrator, adds enums representing movement LimitValues for motors
