+++ Ball9000/src/ArmMain.java
@@ -43,162 +43,11 @@ import lejos.robotics.subsumption.Behavior;
 
 public class ArmMain {
 
-    private static Ball ball;
-/*    private static Vertical vertical;
-    private static Horizontal horizontal;
-    private static Claw claw;
-    private static SenseColour colour;
-    private static Pressure pressure;
-    private static EV3 ev3;
-    private static Port s1;
-    private static Port s2;
-    private static Port s3;*/
-    //todo remove all comments here
+    private static ArmArbitrator armArbitrator;
 
     public static void main(String[] args) {
-        while (Button.ESCAPE.isUp()) { //TODO check that this didn't break anything.
-            ball = new Ball();
-            ball.start();
-/*
-
-            //brick
-            ev3 = (EV3) BrickFinder.getDefault();
-
-            //sensor ports
-            s1 = ev3.getPort("S1");
-            s2 = ev3.getPort("S2");
-            s3 = ev3.getPort("S3");
-
-            //sensors
-            colour = new SenseColour(s3, s1);
-            pressure = new Pressure(s2);
-
-            //motors
-            vertical = new Vertical();
-            horizontal = new Horizontal();
-            claw = new Claw();
-
-
-            // Behaviors
-            Behavior white = new MoveWhite(horizontal, vertical, claw, colour, pressure);
-            Behavior black = new MoveBlack(horizontal, vertical, claw, colour, pressure);
-            Behavior home = new MoveHome(pressure, horizontal, claw, vertical);
-            Behavior wait = new Wait();
-
-
-            // Behavior array
-            Behavior[] bArray = {wait, home, white, black};
-
-            //TODO should the Arbitrator be in its own class (nested class?),
-            // with sensors and motors (or Behavior[]?) as construction parameters?
-
-            // Arbitrator
-            Arbitrator arb = new Arbitrator(bArray);
-            //TODO should Arbitrator and Calibration be implemented
-            // as static nested classes in ArmMain?
-
-
-            //TODO change this to a lejos.utility.TextMenu
-            // implementing choices "Start Arbitrator", "Calibrate", and "Manual Control"
-            boolean cont = true;
-            while (cont) {
-                switch (Button.waitForAnyPress()) {
-                    case (Button.ID_ENTER):
-                        //arb.go();
-                        claw.findZero();
-                        while(colour.getDistance() < 0.7){
-                            vertical.moveArm(-300);
-                        }
-                        vertical.stop();
-                        vertical.resetTacho();
-                        home.action();
-                        break;
-                    case (Button.ID_ESCAPE):
-                        arb.stop();
-                        cont = false;
-                        break;
-                    case (Button.ID_LEFT):
-                        white.action();
-                        break;
-                    case (Button.ID_DOWN):
-                        armTest();
-                        break;
-                    case (Button.ID_UP):
-                        home.action();
-                        break;
-                    case (Button.ID_RIGHT):
-                        clawTest();
-                        break;
-                }
-            }
-            //TODO add app shutdown with ESCAPE button
-        }// while(Button.ESCAPE.isUp())
-        // Exits program when ESCAPE is pressed
-        System.exit(0);
-
-        //TODO add lejos.hardware.Button functions for testing of behaviors separately
-        // i.e. up for Home, left for 90 deg, down for 180 deg, right for calibrate.
-        // center button to activate/deactivate arbitrator?
-
-        //TODO start with arbitrator off, and activate manually?
-
-        //TODO possible to call a behavior directly with Behavior.action?
-
-*/
-        }
-/*
-    private static void armTest() {
-        boolean armTest = true;
-        while (armTest) {
-            switch(Button.waitForAnyPress()) {
-                case (Button.ID_DOWN):
-                    vertical.changeElevation(0);
-                    break;
-                case (Button.ID_UP):
-                    vertical.changeElevation(220);
-                    break;
-                case (Button.ID_ENTER):
-                    while(colour.getDistance() < 0.7){
-                        vertical.moveArm(-300);
-                    }
-                    vertical.stop();
-                    vertical.resetTacho();
-                    break;
-                case (Button.ID_RIGHT):
-                    vertical.changeElevation(110);
-                    break;
-                case (Button.ID_ESCAPE):
-                    armTest = false;
-                    break;
-            }
-        }
-    }
-
-    private static void clawTest() {
-        boolean clawTest = true;
-        while(clawTest) {
-            switch (Button.waitForAnyPress()) {
-                case (Button.ID_UP):
-                    claw.closeClaw();
-                    break;
-                case (Button.ID_DOWN):
-                    claw.openClaw();
-                    break;
-                case (Button.ID_ESCAPE):
-                    clawTest = false;
-                    break;
-                case (Button.ID_ENTER):
-                    claw.findZero();
-                    break;
-            }
-        }
-    }
-    */
-
+            armArbitrator= new ArmArbitrator();
+            armArbitrator.start();
     }//main
-    /*
-    public class Caibrate {} //TODO figure this out
 
-    public class Arbitrate {} //TODO should arbitrator be own class, or nested class?
-    */
-}//class
+}
