--- Ball9000/src/Gyro.java
@@ -1,2 +1,0 @@
-public class Gyro {
-}
